rename to enable-gcp-services
