+++ examples/enable-gcp-services-simple/.expected/diff.patch
@@ -1,6 +1,6 @@
 diff --git a/service_proj1-service-compute.yaml b/service_proj1-service-compute.yaml
 new file mode 100644
-index 0000000..b02d5d8
+index 0000000..6e3a1dd
 --- /dev/null
 +++ b/service_proj1-service-compute.yaml
 @@ -0,0 +1,10 @@
@@ -9,14 +9,14 @@ index 0000000..b02d5d8
 +metadata:
 +  name: proj1-service-compute
 +  annotations:
-+    blueprints.cloud.google.com/managed-by-project-services-list: 'proj1-service'
++    blueprints.cloud.google.com/managed-by-enable-gcp-services: 'proj1-service'
 +spec:
 +  resourceID: compute.googleapis.com
 +  projectRef:
 +    external: proj1
 diff --git a/service_proj1-service-redis.yaml b/service_proj1-service-redis.yaml
 new file mode 100644
-index 0000000..abf3ccb
+index 0000000..83f0c80
 --- /dev/null
 +++ b/service_proj1-service-redis.yaml
 @@ -0,0 +1,10 @@
@@ -25,7 +25,7 @@ index 0000000..abf3ccb
 +metadata:
 +  name: proj1-service-redis
 +  annotations:
-+    blueprints.cloud.google.com/managed-by-project-services-list: 'proj1-service'
++    blueprints.cloud.google.com/managed-by-enable-gcp-services: 'proj1-service'
 +spec:
 +  resourceID: redis.googleapis.com
 +  projectRef:
