+++ examples/enable-gcp-services-simple/README.md
@@ -1,4 +1,4 @@
-# project-service-list: Simple Example
+# enable-gcp-services: Simple Example
 
 ### Overview
 
@@ -9,22 +9,22 @@ In this example, we will see how to generate [project services](https://cloud.go
 Get the example package by running the following commands:
 
 ```shell
-$ kpt pkg get https://github.com/GoogleContainerTools/kpt-functions-catalog.git/examples/project-service-list-simple
+$ kpt pkg get https://github.com/GoogleContainerTools/kpt-functions-catalog.git/examples/enable-gcp-services-simple
 ```
 
 ### Function invocation
 
-Render the configuration, `project-service-list` has been defined declaratively in the Kptfile and configured via `services.yaml`.
+Render the configuration, `enable-gcp-services` has been defined declaratively in the Kptfile and configured via `services.yaml`.
 
 ```shell
-$ kpt fn render project-service-list-simple
+$ kpt fn render enable-gcp-services-simple
 ```
 
 ### Expected result
 
 ```shell
-[RUNNING] "gcr.io/kpt-fn/project-service-list:unstable"
-[PASS] "gcr.io/kpt-fn/project-service-list:unstable"
+[RUNNING] "gcr.io/kpt-fn/enable-gcp-services:unstable"
+[PASS] "gcr.io/kpt-fn/enable-gcp-services:unstable"
   Results:
     [INFO] generated service in object "serviceusage.cnrm.cloud.google.com/v1beta1/Service/proj1-service-compute"
     [INFO] generated service in object "serviceusage.cnrm.cloud.google.com/v1beta1/Service/proj1-service-redis"
