+++ examples/enable-gcp-services-simple/resources.yaml
