+++ functions/go/enable-gcp-services/gcpservices/services_test.go
@@ -1,4 +1,4 @@
-package projectservicelist
+package gcpservices
 
 import (
 	"testing"
