+++ functions/go/enable-gcp-services/metadata.yaml
@@ -1,0 +1,12 @@
+image: gcr.io/kpt-fn/enable-gcp-services
+description: Generate and manage project service enablements for a GCP project.
+tags:
+  - generator
+  - GCP
+  - mutator
+sourceURL: https://github.com/GoogleContainerTools/kpt-functions-catalog/tree/master/functions/go/enable-gcp-services
+examplePackageURLs:
+  - https://github.com/GoogleContainerTools/kpt-functions-catalog/tree/master/examples/enable-gcp-services-simple
+emails:
+  - kpt-team@google.com
+license: Apache-2.0
