--- functions/go/project-service-list/metadata.yaml
@@ -1,12 +1,0 @@
-image: gcr.io/kpt-fn/project-service-list
-description: Generate and manage project service enablements for a GCP project.
-tags:
-  - generator
-  - GCP
-  - mutator
-sourceURL: https://github.com/GoogleContainerTools/kpt-functions-catalog/tree/master/functions/go/project-service-list
-examplePackageURLs:
-  - https://github.com/GoogleContainerTools/kpt-functions-catalog/tree/master/examples/project-service-list-simple
-emails:
-  - kpt-team@google.com
-license: Apache-2.0
