+++ tests/enable-gcp-services/idempotent-prune/.expected/config.yaml
