+++ tests/enable-gcp-services/idempotent-prune/.expected/results.yaml
@@ -4,7 +4,7 @@ metadata:
   name: fnresults
 exitCode: 0
 items:
-  - image: gcr.io/kpt-fn/project-service-list:unstable
+  - image: gcr.io/kpt-fn/enable-gcp-services:unstable
     exitCode: 0
     results:
       - message: pruned service
