+++ tests/enable-gcp-services/idempotent-prune/Kptfile
@@ -1,8 +1,8 @@
 apiVersion: kpt.dev/v1
 kind: Kptfile
 metadata:
-  name: project-service-list-simple
+  name: enable-gcp-services-simple
 pipeline:
   mutators:
-    - image: gcr.io/kpt-fn/project-service-list:unstable
+    - image: gcr.io/kpt-fn/enable-gcp-services:unstable
       configPath: services.yaml
