+++ tests/enable-gcp-services/idempotent-prune/resources.yaml
