+++ tests/enable-gcp-services/idempotent-prune/service_proj1-service-redis.yaml
@@ -3,7 +3,7 @@ kind: Service
 metadata:
   name: proj1-service-redis
   annotations:
-    blueprints.cloud.google.com/managed-by-project-services-list: 'proj1-service'
+    blueprints.cloud.google.com/managed-by-enable-gcp-services: 'proj1-service'
 spec:
   resourceID: redis.googleapis.com
   projectRef:
