+++ tests/enable-gcp-services/idempotent-prune/services.yaml
